feat(csrf-token): Moved auth functions into their own middleware
