+++ server/src/global/auth/authRoutes.ts
@@ -2,16 +2,17 @@
  * Definition of the API routes and invocation of the correct handler function
  */
 import express = require("express");
-const router = express.Router();
-
+import authenticationMiddleware from "../../middleware/authentication";
 import * as authController from "./authController";
+
+const router = express.Router();
 
 router.post("/login", authController.login);
 
 /**
  * =======>>> ALL routes after this point are accessible for loged in users only <<<=======
  */
-router.use(authController.protectRoutes);
+router.use(authenticationMiddleware);
 router.get("/user-data", authController.retrieveUserData);
 router.post("/logout", authController.logout);
